refactor: 데이터 타입 폴더 위치 변경
